applyChanges externalized
