+++ src/test/moveStaticMethodsOutOfClass/index.test.ts
@@ -1,43 +1,16 @@
-import {getAstChanges} from "../../getAstChanges";
-import {Project} from "ts-morph";
 import {assert} from "chai";
-import {AstChangeApplier} from "../../astChangeApplier";
-import {readFileSync} from "fs";
-import {join} from "path";
+import {applyChanges} from "../utilities";
 
 describe('applyClassSplitCommand', () => {
     it('should apply the split command', () => {
-        const sourceFileText1 = readFileSync(join(__dirname, 'sourceFileText1.ts'), 'utf8');
-        const sourceFileText2 = readFileSync(join(__dirname, 'sourceFileText2.ts'), 'utf8');
-        const sourceFileText3 = readFileSync(join(__dirname, 'sourceFileText3.ts'), 'utf8');
+        const {
+            sourceFiles,
+            sourceFileText3,
+        } = applyChanges(__dirname)
 
-        const astChanges = getAstChanges(
-            'index.ts',
-            sourceFileText1,
-            sourceFileText2,
+        assert.equal(
+            sourceFiles[0]![1],
+            sourceFileText3
         );
-
-        const project = new Project({
-            useInMemoryFileSystem: true,
-        });
-        [
-            ['index.ts', sourceFileText2] as const,
-        ].map(
-            ([filePath, sourceFileText]) => {
-                project.createSourceFile(
-                    filePath,
-                    sourceFileText,
-                );
-            }
-        );
-
-        const applier = new AstChangeApplier(
-            project,
-            astChanges,
-        );
-
-        const sourceFiles = applier.applyChanges();
-
-        assert.equal(sourceFiles[0]![1], sourceFileText3);
     });
 });
